refactor: extract into `ReferenceFormatter`
Temporary Code(tm).

Quick and dirty move since it does not really belong in a decorator,
and this move will make it easier to cleanup other code, which will make
it easier to cleanup this code.
