+++ app/decorators/reference_decorator.rb
@@ -1,8 +1,8 @@
-# TODO: Do not cache in database.
-
 class ReferenceDecorator < Draper::Decorator
   delegate_all
+
   delegate :sanitize, to: :helpers
+  delegate :plain_text, :expandable_reference, :expanded_reference, to: :reference_formatter
 
   def link_to_reference
     helpers.link_to keey, helpers.reference_path(self)
@@ -42,36 +42,6 @@ class ReferenceDecorator < Draper::Decorator
     }[reference.review_state]
   end
 
-  # Formats the reference as plaintext (with the exception of <i> tags).
-  def plain_text
-    return generate_plain_text if ENV['NO_REF_CACHE']
-
-    cached = reference.plain_text_cache
-    return cached.html_safe if cached
-
-    reference.set_cache generate_plain_text, :plain_text_cache
-  end
-
-  # Formats the reference with HTML, CSS, etc. Click to show expanded.
-  def expandable_reference
-    return generate_expandable_reference if ENV['NO_REF_CACHE']
-
-    cached = reference.expandable_reference_cache
-    return cached.html_safe if cached
-
-    reference.set_cache generate_expandable_reference, :expandable_reference_cache
-  end
-
-  # Formats the reference with HTML, CSS, etc.
-  def expanded_reference
-    return generate_expanded_reference if ENV['NO_REF_CACHE']
-
-    cached = reference.expanded_reference_cache
-    return cached.html_safe if cached
-
-    reference.set_cache generate_expanded_reference, :expanded_reference_cache
-  end
-
   # TODO: `sanitize(reference.title)` converts ampersands to "&amp;" (only an issue in `Exporters::TaxaAsTxt`).
   # Example: "Brandão &amp; Martins-Neto" from `Taxon.find(429023).authorshiip.reference.decorate.send(:format_plain_text_title)`.
   def format_plain_text_title
@@ -80,56 +50,9 @@ class ReferenceDecorator < Draper::Decorator
 
   private
 
-    def generate_plain_text
-      string = sanitize(reference.author_names_string_with_suffix)
-      string << ' ' unless string.empty?
-      string << sanitize(reference.citation_year) << '. '
-      string << helpers.unitalicize(format_plain_text_title) << ' '
-      string << helpers.add_period_if_necessary(format_plain_text_citation)
-      string
-    end
-
-    def generate_expandable_reference
-      inner_content = []
-      inner_content << generate_expanded_reference
-      inner_content << '[online early]' if reference.online_early?
-      inner_content << format_document_links
-      content = inner_content.reject(&:blank?).join(' ')
-
-      # TODO: `tabindex: 2` is required or tooltips won't stay open even with `data-click-open="true"`.
-      helpers.content_tag :span, sanitize(reference.keey),
-        data: { tooltip: true, allow_html: "true", tooltip_class: "foundation-tooltip" },
-        tabindex: "2", title: content.html_safe
-    end
-
-    def generate_expanded_reference
-      string = sanitize author_names_with_links
-      string << ' ' unless string.empty?
-      string << sanitize(reference.citation_year) << '. '
-      string << format_title_with_link << ' '
-      string << format_italics(helpers.add_period_if_necessary(format_citation))
-      string << ' [online early]' if reference.online_early?
-
-      string
-    end
-
-    # Override in subclasses as necessary.
-    def format_plain_text_citation
-      # `format_citation` + `unitalicize` is to get rid of "*" italics.
-      helpers.unitalicize format_italics(sanitize(format_citation))
-    end
-
-    def author_names_with_links
-      string =  reference.author_names.map do |author_name|
-                  helpers.link_to(sanitize(author_name.name), helpers.author_path(author_name.author))
-                end.join('; ')
-
-      string << sanitize(" #{author_names_suffix}") if author_names_suffix.present?
-      string
-    end
-
-    def format_title_with_link
-      helpers.link_to format_plain_text_title, helpers.reference_path(reference)
+    # TODO: See if what to do here. Added when extracting into `ReferenceFormatter`.
+    def reference_formatter
+      @reference_formatter ||= ReferenceFormatter.new(reference)
     end
 
     def format_italics string
