+++ app/decorators/reference_decorator/book_reference_decorator.rb
@@ -1,9 +1,2 @@
 class BookReferenceDecorator < ReferenceDecorator
-  delegate :publisher, :pagination
-
-  private
-
-    def format_citation
-      sanitize "#{publisher.display_name}, #{pagination}"
-    end
 end
