+++ app/decorators/reference_decorator/missing_reference_decorator.rb
@@ -1,12 +1,4 @@
 class MissingReferenceDecorator < ReferenceDecorator
-  delegate :citation
-
   def format_document_links
   end
-
-  private
-
-    def format_citation
-      sanitize citation
-    end
 end
