+++ app/decorators/reference_decorator/nested_reference_decorator.rb
@@ -1,5 +1,5 @@
 class NestedReferenceDecorator < ReferenceDecorator
-  delegate :pages_in, :nesting_reference
+  delegate :nesting_reference
 
   # Fall back to nesting reference's PDF is nestee does not have one.
   def pdf_link
@@ -8,14 +8,6 @@ class NestedReferenceDecorator < ReferenceDecorator
   end
 
   private
-
-    def format_plain_text_citation
-      sanitize "#{pages_in} #{nesting_reference.decorate.plain_text}"
-    end
-
-    def format_citation
-      sanitize "#{pages_in} #{sanitize nesting_reference.decorate.expanded_reference}"
-    end
 
     def nesting_reference_pdf_link
       return unless nesting_reference.downloadable?
