+++ app/decorators/reference_decorator/unknown_reference_decorator.rb
@@ -1,9 +1,2 @@
 class UnknownReferenceDecorator < ReferenceDecorator
-  delegate :citation
-
-  private
-
-    def format_citation
-      sanitize citation
-    end
 end
