+++ spec/decorators/reference_decorator_spec.rb
@@ -35,73 +35,6 @@ describe ReferenceDecorator do
     end
   end
 
-  describe "#plain_text" do
-    context "with unsafe tags" do
-      let!(:author_names) { [create(:author_name, name: 'Ward, P. S.')] }
-      let(:reference) do
-        create :unknown_reference, author_names: author_names,
-          citation_year: "1874", title: "Les fourmis de la Suisse.", citation: '32 pp.'
-      end
-
-      it "escapes everything, buts let italics through" do
-        reference.author_names = [create(:author_name, name: '<script>')]
-        expect(reference.decorate.plain_text).
-          to eq '1874. Les fourmis de la Suisse. 32 pp.'
-      end
-
-      it "escapes the citation year" do
-        reference.update!(citation_year: '<script>')
-        expect(reference.decorate.plain_text).
-          to eq 'Ward, P. S. . Les fourmis de la Suisse. 32 pp.'
-      end
-
-      it "escapes the title" do
-        reference.update!(title: '*foo*<script>')
-        expect(reference.decorate.plain_text).to eq 'Ward, P. S. 1874. foo. 32 pp.'
-      end
-    end
-  end
-
-  describe "#expandable_reference and #expanded_reference" do
-    context 'with unsafe tags' do
-      let!(:reference) do
-        author_name = create :author_name, name: '<script>xss</script>'
-        journal = create :journal, name: '<script>xss</script>'
-        create :article_reference, journal: journal, pagination: '<script>xss</script>',
-          series_volume_issue: '<script>xss</script>', title: '<script>xss</script>',
-          citation_year: '<script>xss</script>', author_names: [author_name],
-          author_names_suffix: '<script>xss</script>', date: '<script>xss</script>'
-      end
-
-      describe '#expandable_reference' do
-        it "sanitizes them" do
-          results = reference.decorate.expandable_reference
-          expect(results).to_not include '<script>xss</script>'
-
-          # NOTE: `.to include`, not `.to_not include` since it's wrapped inside the title attribute of the tooltip.
-          expect(results).to include '&lt;script&gt;xss&lt;/script&gt;'
-          expect(results).to include 'xss'
-        end
-      end
-
-      describe '#expanded_reference' do
-        it "sanitizes them" do
-          results = reference.decorate.expanded_reference
-          expect(results).to_not include '<script>xss</script>'
-          expect(results).to_not include '&lt;script&gt;xss&lt;/script&gt;'
-          expect(results).to include 'xss'
-        end
-      end
-    end
-
-    context 'when reference is online early' do
-      let(:reference) { create :article_reference, online_early: true }
-
-      specify { expect(reference.decorate.expandable_reference).to include ' [online early]' }
-      specify { expect(reference.decorate.expanded_reference).to include ' [online early]' }
-    end
-  end
-
   describe "#format_review_state" do
     let(:reference) { build_stubbed :article_reference }
